Convert encryption and chunk storage to filehandle-based versions (Kostas Chatzikokolakis).
git-svn-id: http://brackup.googlecode.com/svn/trunk@281 c3152786-d43a-0410-8499-09eb43983aed

+++ lib/Brackup/Util.pm
@@ -5,11 +5,12 @@ require Exporter;
 
 use vars qw(@ISA @EXPORT_OK);
 @ISA = ('Exporter');
-@EXPORT_OK = qw(tempfile tempdir slurp valid_params noclobber_filename io_print_to_fh);
+@EXPORT_OK = qw(tempfile tempdir slurp valid_params noclobber_filename io_print_to_fh io_sha1);
 
 use File::Path qw();
 use Carp;
 use Fcntl qw(O_RDONLY);
+use Digest::SHA1;
 
 my $mainpid = $$;
 my $_temp_directory;
@@ -17,7 +18,7 @@ my $_temp_directory;
 END {
     # will happen after File::Temp's cleanup
     if ($$ == $mainpid and $_temp_directory) {
-        File::Path::rmtree($_temp_directory, 0, 1);
+        File::Path::rmtree($_temp_directory, 0, 1) unless $ENV{BRACKUP_TEST_NOCLEANUP};
     }
 }
 use File::Temp ();
@@ -26,7 +27,7 @@ sub _get_temp_directory {
     # Create temporary directory if we need one. By default, all temporary
     # files will be placed in it.
     unless (defined($_temp_directory)) {
-        $_temp_directory = File::Temp::tempdir(CLEANUP => 1);
+        $_temp_directory = File::Temp::tempdir(CLEANUP => $ENV{BRACKUP_TEST_NOCLEANUP} ? 0 : 1);
     }
 
     return $_temp_directory;
@@ -69,14 +70,31 @@ sub noclobber_filename {
 
 # Prints all data from an IO::Handle to a filehandle
 sub io_print_to_fh {
-    my ($io_handle, $fh) = @_;
+    my ($io_handle, $fh, $sha1) = @_;
     my $buf;
     my $bytes = 0;
+
     while($io_handle->read($buf, 4096)) {
         print $fh $buf;
         $bytes += length $buf;
+        $sha1->add($buf) if $sha1;
     }
+
     return $bytes;
+}
+
+# computes sha1 of data in an IO::Handle
+sub io_sha1 {
+    my ($io_handle) = @_;
+    
+    my $sha1 = Digest::SHA1->new;
+    my $buf;
+    
+    while($io_handle->read($buf, 4096)) {
+        $sha1->add($buf);
+    }
+
+    return $sha1->hexdigest;
 }
 
 1;
